PE-80 | Token Generation Demo
Also changed Integers to Doubles in user ratings lists

+++ src/main/java/com/ProjectEureka/backend/models/User.java
@@ -36,7 +36,7 @@ public class User {
 
     private int role;
 
-    private ArrayList<Integer> ratings;
+    private ArrayList<Double> ratings;
 
     @Field("average_rating")
     private double averageRating;
